+++ python/cudf/dataframe/series.py
@@ -1849,11 +1849,37 @@ class Series(object):
         DataFrame
 
         """
+
+        if isinstance(q, Number) or utils.is_list_like(q):
+            np_array_q = np.asarray(q)
+            if np.logical_or(np_array_q < 0, np_array_q > 1).any():
+                raise ValueError("percentiles should all \
+                             be in the interval [0, 1]")
+
+        # Beyond this point, q either being scalar or list-like
+        # will only have values in range [0, 1]
+
+        if isinstance(q, Number):
+            res = self._column.quantile(q, interpolation, exact)
+            if len(res) == 0:
+                return np.nan
+            else:
+                # if q is an int/float, we shouldn't be constructing series
+                return res.pop()
+
         if not quant_index:
             return Series(self._column.quantile(q, interpolation, exact))
         else:
-            return Series(self._column.quantile(q, interpolation, exact),
-                          index=as_index(np.asarray(q)))
+            from cudf.dataframe.columnops import column_empty_like
+            np_array_q = np.asarray(q)
+            if len(self) == 0:
+                result = column_empty_like(np_array_q, dtype=self.dtype,
+                                           masked=True,
+                                           newsize=len(np_array_q))
+            else:
+                result = self._column.quantile(q, interpolation, exact)
+            return Series(result,
+                          index=as_index(np_array_q))
 
     def describe(self, percentiles=None, include=None, exclude=None):
         """Compute summary statistics of a Series. For numeric